trading_engine/requirements.txt
fastapi
uvicorn
pandas
requests
psycopg2-binary
sqlalchemy
pydantic
alpaca-trade-api
pytz
alpaca-py
aiohttp